pyproject.toml
[project]
name = "raw-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.39",
    "celery>=5.6.2",
    "django>=5.2.10",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "exifread>=3.5.1",
    "imageio>=2.37.2",
    "pillow>=12.1.0",
    "rawpy==0.24.0",
    "redis>=7.1.0",
]
